requirements.txt
feedparser
aiohttp
trafilatura
zhconv
deep-translator
beautifulsoup4
python-dotenv
